+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-21T15:10:16.389Z",
+    "2026-02-23T02:47:34.283Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {